Updating conftest.py to use updated module path.
